+++ docker/mcrit/entry_worker.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
 set -eu
+# shellcheck source-path=SCRIPTDIR source=entry_common.sh
+. /entry_common.sh
 cd /opt/mcrit
 exec python -m mcrit worker